Cooler branch (#33)
* feat(routes) users routes created and moved it to a new folder: routes

* bug(backend - sequelize)-cannot get any data from database through router

* bugfix(backend - sequelize) figured out that db was overwriting seed file

* feat(user model)added validation for several columns

* fix(backend routes)fixed user put route

* feat(backend routes and models)seeded products and set api/products path

* feat(routes) added reviews routes, updated seed file for reviews, commented out route for accessing all reviews for a product within product route file (now in review route file)

* feat(routes) work in process on transactions route

* feat(routes) no really, this time transactions is done

* feat(routes) order route started

* feat(routes) updated orders routes for cart. Includes get and post, and also commented out put route for purchasing (depends on other branch to be merged to master)

* feat(routes) finished draft of orders and transactions - to test

* pushing to pull other branch

* feat(routes) mostly added categories routes

* pushing to pull

* feat(routes and seed file) updated

* feat(routes) now with more delete categories from a product

* feat(routes) done! still adding comments

+++ db/models/transaction.js
@@ -6,7 +6,7 @@ const Sequelize = require('sequelize')
 const db = require('APP/db')
 
 const Transaction = db.define('transactions', {
-    sellingPrice: { type: Sequelize.DECIMAL(10, 2) },
+    sellingPrice: { type: Sequelize.INTEGER },
     quantity: { type: Sequelize.INTEGER, defaultValue: 1 }
 }, {})
